requirements.txt
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.74
langchain-openai==0.3.30
langchain-text-splitters==0.3.9
openai==1.100.2
python-dotenv==1.1.1
fastapi==0.116.1
langserve==0.3.1
sse_starlette==3.0.2
bs4==0.0.2
faiss-cpu==1.12.0.post1
tqdm==4.67.1
tiktoken==0.11.0
langchain_chroma==0.2.5
langchainhub==0.1.21